Simplified upload code for point light data. Removed warning when using UNITY_MATRIX_MVP.

+++ com.unity.render-pipelines.universal/Shaders/Utils/Deferred.hlsl
@@ -5,7 +5,7 @@ struct PointLightData
 {
     float3 WsPos;
     float Radius;
-    float3 Color;
+    half3 Color;
 };
